feat(app): 基本画线功能
