+++ public/index.html
@@ -13,6 +13,4 @@
     </div>
   </div>
 </body>
-<script src="./SvgDom.js"></script>
-<script src="./index.js"></script>
 </html>
